require cycle warning removed by extracting dbcontext to its own file
